extract/transformとpages(可視化・分析)を疎結合にするリポジトリ再編
- data取得(extract)と生データ加工(transform)を束ねるpipelines層を実装し、
  未使用だったPipeline基底クラスの下にSuumoPipeline/MansionReviewPipelineを追加。
  load.pyはpipelines.registryを呼ぶだけの薄いCLIにした。
- pages(streamlit)が要求する分析データのスキーマをanalysis/schema.pyに切り出し、
  Pipeline.validateはそこを参照する形にして、可視化側の要件がpipeline実装に
  直書きされないようにした。
- pages側で重複していたCSV一覧取得/読み込み/色計算/範囲フィルタをanalysis/配下
  (loader.py, colors.py, filters.py)に共通化し、各pageはそれらを呼ぶだけにした。
- ルート直下にあった transform.py は同名の transform/ パッケージに隠蔽されて
  import不能な死んだコードだったため、transform/dedup.py として統合。
- extract/utils.py にあった生データ比較処理(get_diff_records)をtransform/diff.py
  に移動し、IF層とデータ加工層の責務を分離。
- plateau.py/price3d.pyのハードコードされた存在しないパス/カラム名を、他ページと
  同じdata/analytics経由の読み込みに修正。extract/mansionreview_history.pyの
  誤ったimportも修正。

streamlit.testing.v1.AppTestで各pageがサンプルデータに対しエラーなく描画される
ことを確認済み(analytics.pyのplotly_express未インストールは既存の別課題として残置)。
