+++ README.md
@@ -21,16 +21,18 @@ uv run streamlit run Home.py
 
 `Home.py` がメイン画面で、`pages/` 配下の各ページ(analytics, app_estate, plateau, price3d)がサイドバーから選択できます。
 
-リポジトリ構造を考える
-- scrap
-  - suumo.py
-  - mansionreview.py
-  - mansionreview history
-- data
-  - ActiveListings
-    - suumoyyyymmdd.csv
-    - mansionreview yyyymmd.csv
-  - TransactionHistory
-    - くらまえ.csv
-- analytics
-  - streamlit
+# リポジトリ構造
+
+データの取得・加工(IF+生データ加工)と、可視化・分析(streamlit)を疎結合にする方針で構成しています。
+可視化側は`data/analytics/`配下のCSVというインターフェース越しにのみデータ取得・加工側とつながっており、
+互いの実装の詳細(スクレイピング方法やDataFrameの中間表現など)には依存しません。
+
+- `extract/` … 外部サイト・外部API(ジオコーディング等)へのアクセスのみを担うIF層
+- `transform/` … `extract`が返す生データを分析用データに加工する層(重複除去・ジオコーディング付与など)
+- `pipelines/` … サイトごとに`extract`+`transform`をまとめて実行するオーケストレーション層(`Pipeline`基底クラス)
+- `load.py` … `pipelines`を呼び出すCLIエントリポイント。`data/rawdata/`に生データ、`data/analytics/`に加工済みデータを保存する
+- `analysis/` … 可視化・分析(pages)側が共通で使うロジック(CSV読み込み、絞り込み、色計算)と、分析データが満たすべきスキーマ(`analysis/schema.py`)
+- `pages/` … streamlitの各画面。`analysis/`経由で`data/analytics/`配下のデータを読み込んで可視化するだけの薄い層
+- `data/`
+  - `rawdata/` … `extract`直後の生データ
+  - `analytics/` … `transform`後の分析用データ(pagesはここだけを参照する)
