+++ pages/plateau.py
@@ -5,11 +5,18 @@ import pandas as pd
 from geopandas import GeoDataFrame
 import requests
 
-from utils import scale_color
+from analysis.colors import scale_color
+from analysis.loader import list_csv_files, load_csv
+
+target_folder = "activelist"
 
 # 都市データ読み込み（例: 渋谷区）
 dataset = load_dataset("plateau-13106-taito-ku-2023")
-dataframe=pd.read_csv("data/activelist/mansionreview_20250714.csv",index_col=0)
+
+with st.sidebar:
+    base_data_name=st.selectbox('対象のデータ', list_csv_files(target_folder))
+
+dataframe=load_csv(target_folder, base_data_name)
 # Apply the function to create a color column
 dataframe['color'] = dataframe['坪単価'].apply(lambda x: scale_color(x))
 
@@ -58,7 +65,6 @@ building3d = pdk.Layer(
 )
 
 # scrapデータの処理
-dataframe["name"] = dataframe["建物名"]
 data = dataframe.to_dict(orient='records')
 st.session_state.candidates=pd.DataFrame(columns=dataframe.columns)
