+++ pages/price3d.py
@@ -1,16 +1,22 @@
 import streamlit as st
-import pandas as pd
 import pydeck as pdk
 import numpy as np
 from cmcrameri import cm
 
+from analysis.loader import list_csv_files, load_csv
+
+target_folder = "activelist"
+
+with st.sidebar:
+    base_data_name=st.selectbox('対象のデータ', list_csv_files(target_folder))
+
 # 元のデータフレームに追加します
-data=pd.read_csv("suumo_20250714.csv",index_col=0)
+data=load_csv(target_folder, base_data_name)
 
 colors = cm.hawaii_r(np.linspace(0, 1, 256))
 rgb_colors = (colors[:, :3] * 255).astype(int).tolist()
 
-data["log"]=data["price per unit area"].apply(lambda x: np.log10(x))
+data["log"]=data["坪単価"].apply(lambda x: np.log10(x))
 min_height = data["log"].min()
 max_height = data["log"].max()
 data["color"] = data["log"].apply(
@@ -22,7 +28,7 @@ exclude_wards = ["千代田区", "中央区", "港区", "新宿区", "文京区"
 # 都心3区のレコードを除外
 data = data[~data["address"].str.contains('|'.join(exclude_wards))]
 
-data["view"]=data["price per unit area"]
+data["view"]=data["坪単価"]
 
 # PyDeckを使用して地図を描画します
 st.pydeck_chart(pdk.Deck(
