+++ pipelines/core/pipeline.py
@@ -1,11 +1,18 @@
 from abc import ABC, abstractmethod
 
+from analysis.schema import REQUIRED_COLUMNS
+
+# extract直後の生データに最低限必要なカラム(サイトによらず共通)
+RAW_REQUIRED_COLUMNS = ["name", "price", "address", "area", "age"]
+
 
 class Pipeline(ABC):
-    """extract/transformを行うパイプラインの基底クラス"""
+    """extract/transformを行うパイプラインの基底クラス。
 
-    # streamlit側(pages/app_estate.py)の表示に必要なカラム
-    REQUIRED_COLUMNS = ["address", "area", "age", "lons", "lats", "坪単価"]
+    IF(extract)と生データ加工(transform)を1サイト単位でまとめる役割のみを持ち、
+    可視化・分析側(analysis, pages)の実装には関与しない。
+    可視化側が要求するデータ契約はanalysis.schemaを参照する。
+    """
 
     @abstractmethod
     def extract(self):
@@ -15,15 +22,19 @@ class Pipeline(ABC):
     def transform(self, data):
         raise NotImplementedError
 
+    def validate_raw(self, data):
+        """extract直後の生データが最低限の形になっているか検証する"""
+        return all(col in data for col in RAW_REQUIRED_COLUMNS)
+
     def validate(self, data):
-        """transform後のdfがstreamlitで読み込める形になっているか検証する"""
+        """transform後のdfがanalysis.schemaの要件を満たしているか検証する"""
         if data.empty:
             return False
 
-        if not all(col in data.columns for col in self.REQUIRED_COLUMNS):
+        if not all(col in data.columns for col in REQUIRED_COLUMNS):
             return False
 
-        if data[self.REQUIRED_COLUMNS].isnull().any().any():
+        if data[REQUIRED_COLUMNS].isnull().any().any():
             return False
 
         return True
