+++ pipelines/registry.py
@@ -1,0 +1,9 @@
+"""サイト名からPipelineを引くレジストリ。load.py(CLI)から使う。"""
+
+from .mansionreview import MansionReviewPipeline
+from .suumo import SuumoPipeline
+
+PIPELINES = {
+    "suumo": SuumoPipeline,
+    "mansionreview": MansionReviewPipeline,
+}
