+++ transform/suumo.py
@@ -1,4 +1,4 @@
-from .utils import get_lat_lon
+from .geocode import get_lat_lon
 
 def transform(data):
